+++ src/main/java/net/semanticmetadata/lire/solr/indexing/SimpleBackslashReplacer.java
@@ -1,0 +1,21 @@
+package net.semanticmetadata.lire.solr.indexing;
+
+/**
+ * Created by mlux on 19.12.2016.
+ */
+public class SimpleBackslashReplacer implements ImageDataProcessor {
+    @Override
+    public CharSequence getTitle(String filename) {
+        return filename.replaceAll("\\\\", "/");
+    }
+
+    @Override
+    public CharSequence getIdentifier(String filename) {
+        return filename.replaceAll("\\\\", "/");
+    }
+
+    @Override
+    public CharSequence getAdditionalFields(String filename) {
+        return "";
+    }
+}